feat(ClipboardModule): move providers to `forRoot()` static method
ClipboardModule now has `forRoot()` static method that provide the service. This is to make sure
that there are only one instance of ClipboardService in the application.

BREAKING CHANGE: ClipboardModule no longer provides ClipboardService. Import
ClipboardModule.forRoot() at the root module of the application and import just ClipboardModule for
the rest.